[test]:(data-source-management) Add unit tests
