+++ packages/base/src/hooks/useTaskSource/index.tsx
@@ -1,11 +1,8 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { useBoolean } from 'ahooks';
 import { Select } from 'antd';
 import { ResponseCode } from '@actiontech/shared/lib/enum';
-import {
-  IDBServiceSyncTaskTip,
-  IDatabaseSource
-} from '@actiontech/shared/lib/api/base/service/common';
+import { IDBServiceSyncTaskTip } from '@actiontech/shared/lib/api/base/service/common';
 import { useDbServiceDriver } from '@actiontech/shared/lib/global';
 import { DatabaseTypeLogo } from '@actiontech/shared';
 import DBServiceSyncTaskService from '@actiontech/shared/lib/api/base/service/DBServiceSyncTask';
