+++ packages/base/src/page/Nav/SideMenu/UserMenu/components/__tests__/UserNavigate.ce.test.tsx
