+++ packages/base/src/page/SyncDataSource/AddPage/index.test.tsx
@@ -4,19 +4,15 @@ import {
   getAllBySelector,
   getBySelector
 } from '@actiontech/shared/lib/testUtil/customQuery';
-
 import syncTaskList from '../../../testUtils/mockApi/syncTaskList';
 import ruleTemplate from 'sqle/src/testUtils/mockApi/rule_template';
 import { mockUseCurrentProject } from '@actiontech/shared/lib/testUtil/mockHook/mockUseCurrentProject';
-import { mockProjectInfo } from '@actiontech/shared/lib/testUtil/mockHook/data';
 import EmitterKey from '../../../data/EmitterKey';
 import EventEmitter from '../../../utils/EventEmitter';
 
 import AddSyncTask from '.';
 
 describe('page/SyncDataSource/AddPage', () => {
-  const projectID = mockProjectInfo.projectID;
-
   const customRender = () => {
     return superRender(<AddSyncTask />);
   };
@@ -36,88 +32,49 @@ describe('page/SyncDataSource/AddPage', () => {
 
   it('render add sync task snap', async () => {
     const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(1000));
-
-    expect(screen.getByText('返回同步任务列表')).toBeInTheDocument();
-    expect(screen.getByText('重 置')).toBeInTheDocument();
-    expect(screen.getByText('提 交')).toBeInTheDocument();
-
-    expect(screen.getByText('添加同步任务')).toBeInTheDocument();
-    expect(screen.getByText('基础配置')).toBeInTheDocument();
-    expect(screen.getByText('SQL审核配置')).toBeInTheDocument();
-    expect(screen.getByText('自定义任务同步周期')).toBeInTheDocument();
+    await act(async () => jest.advanceTimersByTime(3000));
 
     expect(baseElement).toMatchSnapshot();
   });
 
-  it('render click list a link', async () => {
-    const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(1000));
-
-    const aLink = getBySelector(
-      '.actiontech-page-header-namespace a',
-      baseElement
-    );
-    expect(aLink).toHaveAttribute(
-      'href',
-      `/project/${projectID}/sync-data-source`
-    );
-  });
-
   it('render reset form cont', async () => {
     const eventEmitSpy = jest.spyOn(EventEmitter, 'emit');
-    const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(1000));
-
-    // version
-    await act(async () => {
-      fireEvent.change(getBySelector('#version', baseElement), {
-        target: {
-          value: '5.23.04.0'
-        }
-      });
-      await act(async () => jest.advanceTimersByTime(300));
-    });
-    expect(getBySelector('#version', baseElement)).toHaveAttribute(
-      'value',
-      '5.23.04.0'
-    );
+    customRender();
+    await act(async () => jest.advanceTimersByTime(3000));
 
     await act(async () => {
       fireEvent.click(screen.getByText('重 置'));
-      await act(async () => jest.advanceTimersByTime(300));
+      await act(async () => jest.advanceTimersByTime(0));
     });
     expect(eventEmitSpy).toHaveBeenCalledWith(
       EmitterKey.DMS_SYNC_TASK_RESET_FORM
     );
-    await act(async () => jest.advanceTimersByTime(300));
   });
 
   it('render form item for prepare api', async () => {
     const requestRuleGlobal = ruleTemplate.getRuleTemplateTips();
-    const requestRuleProject = ruleTemplate.getProjectRuleTemplateTips();
     const requestTaskSourceListTips = syncTaskList.getTaskSourceListTips();
     const { baseElement } = customRender();
 
-    await act(async () => jest.advanceTimersByTime(3300));
-    expect(requestTaskSourceListTips).toHaveBeenCalled();
+    expect(requestTaskSourceListTips).toHaveBeenCalledTimes(1);
+    expect(requestRuleGlobal).toHaveBeenCalledTimes(1);
+
+    await act(async () => jest.advanceTimersByTime(3000));
 
     // source
     fireEvent.mouseDown(getBySelector('#source', baseElement));
-    await act(async () => jest.advanceTimersByTime(300));
     fireEvent.click(getBySelector('div[title="source1"]', baseElement));
-    await act(async () => jest.advanceTimersByTime(300));
-    expect(requestTaskSourceListTips).toHaveBeenCalledWith({
-      project_uid: projectID
-    });
+
+    await act(async () => jest.advanceTimersByTime(3000));
+
+    expect(screen.getByText('动态字段1')).toBeInTheDocument();
 
     // instanceType
     fireEvent.mouseDown(getBySelector('#instanceType', baseElement));
     await act(async () => jest.advanceTimersByTime(300));
     fireEvent.click(getBySelector('span[title="mysql"]', baseElement));
     await act(async () => jest.advanceTimersByTime(6300));
-    expect(requestRuleGlobal).toHaveBeenCalled();
-    expect(requestRuleProject).toHaveBeenCalled();
+    expect(requestRuleGlobal).toHaveBeenCalledTimes(2);
 
     // ant-select-clear
     const clearIcon = getAllBySelector('.ant-select-clear', baseElement);
@@ -130,7 +87,7 @@ describe('page/SyncDataSource/AddPage', () => {
   it('render add submit for success', async () => {
     const requestSubmit = syncTaskList.addTaskSource();
     const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(3300));
+    await act(async () => jest.advanceTimersByTime(3000));
 
     // name
     fireEvent.change(getBySelector('#name', baseElement), {
@@ -138,21 +95,19 @@ describe('page/SyncDataSource/AddPage', () => {
         value: 'name-sync-source'
       }
     });
-    await act(async () => jest.advanceTimersByTime(300));
 
     // source
     fireEvent.mouseDown(getBySelector('#source', baseElement));
-    await act(async () => jest.advanceTimersByTime(300));
+
     fireEvent.click(getBySelector('div[title="source1"]', baseElement));
     await act(async () => jest.advanceTimersByTime(300));
 
-    // version
-    fireEvent.change(getBySelector('#version', baseElement), {
-      target: {
-        value: '5.23.04.0'
-      }
+    fireEvent.change(getBySelector('#params_key1', baseElement), {
+      target: { value: 'param1' }
     });
-    await act(async () => jest.advanceTimersByTime(300));
+    fireEvent.change(getBySelector('#params_key2', baseElement), {
+      target: { value: 111 }
+    });
 
     // url
     fireEvent.change(getBySelector('#url', baseElement), {
@@ -168,12 +123,23 @@ describe('page/SyncDataSource/AddPage', () => {
     fireEvent.click(getBySelector('span[title="mysql"]', baseElement));
 
     // ruleTemplateName
-    await act(async () => jest.advanceTimersByTime(6300));
+    await act(async () => jest.advanceTimersByTime(3000));
     fireEvent.mouseDown(getBySelector('#ruleTemplateName', baseElement));
     await act(async () => jest.advanceTimersByTime(300));
     fireEvent.click(
       getBySelector('div[title="custom_template_b"]', baseElement)
     );
+
+    // sql query config
+    fireEvent.click(getBySelector('#needAuditForSqlQuery', baseElement));
+
+    fireEvent.mouseDown(
+      getBySelector('#allowQueryWhenLessThanAuditLevel', baseElement)
+    );
+    await act(async () => jest.advanceTimersByTime(300));
+
+    fireEvent.click(getBySelector('div[title="notice"]', baseElement));
+
     await act(async () => jest.advanceTimersByTime(300));
 
     fireEvent.click(screen.getByText('提 交'));
@@ -183,24 +149,30 @@ describe('page/SyncDataSource/AddPage', () => {
     await act(async () => jest.advanceTimersByTime(3000));
     expect(requestSubmit).toHaveBeenCalled();
     expect(requestSubmit).toHaveBeenCalledWith({
-      database_source_service: {
+      db_service_sync_task: {
         name: 'name-sync-source',
         db_type: 'mysql',
         source: 'source1',
         sqle_config: {
           rule_template_id: '2',
-          rule_template_name: 'custom_template_b'
+          rule_template_name: 'custom_template_b',
+          sql_query_config: {
+            audit_enabled: true,
+            allow_query_when_less_than_audit_level: 'notice'
+          }
         },
         cron_express: '0 0 * * *',
         url: 'http://192.168.1.1:27601',
-        version: '5.23.04.0'
-      },
-      project_uid: projectID
+        additional_params: [
+          { key: 'key1', value: 'param1' },
+          { key: 'key2', value: '111' }
+        ]
+      }
     });
     await act(async () => jest.advanceTimersByTime(300));
     expect(screen.getByText('添加同步任务成功')).toBeInTheDocument();
     expect(
-      screen.getByText('到同步任务列表查看看看添加的同步任务')
+      screen.getByText('到同步任务列表查看添加的同步任务')
     ).toBeInTheDocument();
     expect(baseElement).toMatchSnapshot();
     fireEvent.click(screen.getByText('关 闭'));
