+++ packages/base/src/page/SyncDataSource/Form/index.test.tsx
@@ -1,18 +1,17 @@
-import { act, cleanup, fireEvent, screen } from '@testing-library/react';
+import { act, cleanup, fireEvent } from '@testing-library/react';
 import { superRender } from '../../../testUtils/customRender';
 import { getBySelector } from '@actiontech/shared/lib/testUtil/customQuery';
-
 import { Form } from 'antd';
 import { IListDatabaseSourceService } from '@actiontech/shared/lib/api/base/service/common';
-import SyncTaskForm, { SyncTaskFormFields } from '.';
+import SyncTaskForm from '.';
 import { renderHooksWithTheme } from '@actiontech/shared/lib/testUtil/customRender';
-import { mockProjectInfo } from '@actiontech/shared/lib/testUtil/mockHook/data';
 import syncTaskList from '../../../testUtils/mockApi/syncTaskList';
 import ruleTemplate from 'sqle/src/testUtils/mockApi/rule_template';
+import useTaskSource from '../../../hooks/useTaskSource';
+import { mockUseDbServiceDriver } from '@actiontech/shared/lib/testUtil/mockHook/mockUseDbServiceDriver';
+import { SyncTaskFormFields } from './index.type';
 
 describe('page/SyncDataSource/SyncTaskForm', () => {
-  const projectName = mockProjectInfo.projectName;
-
   const customRender = (params?: {
     loading?: boolean;
     defaultValue?: IListDatabaseSourceService;
@@ -20,21 +19,26 @@ describe('page/SyncDataSource/SyncTaskForm', () => {
     const { result } = renderHooksWithTheme(() =>
       Form.useForm<SyncTaskFormFields>()
     );
+    const { result: taskSourceTips } = renderHooksWithTheme(() =>
+      useTaskSource()
+    );
     const isLoading = params?.loading ?? false;
     const defaultData = params?.defaultValue ?? undefined;
+
     return superRender(
       <SyncTaskForm
         form={result.current[0]}
         defaultValue={defaultData}
         loading={isLoading}
         title="这里是标题"
-        projectName={projectName}
+        taskSourceTips={taskSourceTips.current}
       />
     );
   };
 
   beforeEach(() => {
     jest.useFakeTimers();
+    mockUseDbServiceDriver();
     syncTaskList.mockAllApi();
     ruleTemplate.mockAllApi();
   });
@@ -47,13 +51,13 @@ describe('page/SyncDataSource/SyncTaskForm', () => {
 
   it('render for loading form', async () => {
     const { baseElement } = customRender({ loading: true });
-    await act(async () => jest.advanceTimersByTime(1000));
+    await act(async () => jest.advanceTimersByTime(3000));
     expect(baseElement).toMatchSnapshot();
   });
 
   it('render name rule', async () => {
     const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(9300));
+    await act(async () => jest.advanceTimersByTime(3000));
 
     const nameEle = getBySelector('#name', baseElement);
     await act(async () => {
@@ -77,60 +81,10 @@ describe('page/SyncDataSource/SyncTaskForm', () => {
   });
 
   it('render for prepare api request', async () => {
-    const requestTaskSourceList = syncTaskList.getTaskSourceListTips();
     const requestTemplateTips = ruleTemplate.getRuleTemplateTips();
-    const requestTemplateProject = ruleTemplate.getProjectRuleTemplateTips();
-    const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(9300));
-    expect(requestTaskSourceList).toHaveBeenCalled();
-    expect(requestTemplateTips).toHaveBeenCalled();
-    expect(requestTemplateProject).toHaveBeenCalled();
-    expect(baseElement).toMatchSnapshot();
-  });
-
-  it('render source form item', async () => {
-    const requestTaskSourceList = syncTaskList.getTaskSourceListTips();
-    const { baseElement } = customRender();
-
-    await act(async () => jest.advanceTimersByTime(300));
-    const typeEle = getBySelector('#source');
-    fireEvent.mouseDown(typeEle, baseElement);
+    customRender();
     await act(async () => jest.advanceTimersByTime(3000));
-    expect(requestTaskSourceList).toHaveBeenCalled();
-    fireEvent.click(getBySelector('div[title="source1"]', baseElement));
-    await act(async () => jest.advanceTimersByTime(300));
-    expect(baseElement).toMatchSnapshot();
-  });
-
-  it('render dbType form item', async () => {
-    const requestTaskSourceList = syncTaskList.getTaskSourceListTips();
-    const requestTemplateTips = ruleTemplate.getRuleTemplateTips();
-    const requestTemplateProject = ruleTemplate.getProjectRuleTemplateTips();
-    const { baseElement } = customRender();
-
-    const sourceEle = getBySelector('#source');
-    fireEvent.mouseDown(sourceEle, baseElement);
-    await act(async () => jest.advanceTimersByTime(3000));
-    expect(requestTaskSourceList).toHaveBeenCalled();
-    fireEvent.click(getBySelector('div[title="source1"]', baseElement));
-    await act(async () => jest.advanceTimersByTime(300));
-
-    await act(async () => jest.advanceTimersByTime(300));
-    const typeEle = getBySelector('#instanceType');
-    fireEvent.mouseDown(typeEle, baseElement);
-    await act(async () => jest.advanceTimersByTime(300));
-    fireEvent.click(getBySelector('span[title="mysql"]', baseElement));
-    await act(async () => jest.advanceTimersByTime(300));
-    expect(baseElement).toMatchSnapshot();
-
-    await act(async () => jest.advanceTimersByTime(6300));
     expect(requestTemplateTips).toHaveBeenCalled();
-    expect(requestTemplateProject).toHaveBeenCalled();
-
-    const ruleTemplateEle = getBySelector('#ruleTemplateName');
-    fireEvent.mouseDown(ruleTemplateEle, baseElement);
-    await act(async () => jest.advanceTimersByTime(300));
-    expect(baseElement).toMatchSnapshot();
   });
 
   it('render syncInterval for form item', async () => {
