+++ packages/base/src/page/SyncDataSource/UpdatePage/index.test.tsx
@@ -1,18 +1,15 @@
 import Router, { useNavigate } from 'react-router-dom';
 import { act, cleanup, fireEvent, screen } from '@testing-library/react';
 import { superRender } from '../../../testUtils/customRender';
-
 import syncTaskList from '../../../testUtils/mockApi/syncTaskList';
 import ruleTemplate from 'sqle/src/testUtils/mockApi/rule_template';
-import { mockUseCurrentProject } from '@actiontech/shared/lib/testUtil/mockHook/mockUseCurrentProject';
-import { mockProjectInfo } from '@actiontech/shared/lib/testUtil/mockHook/data';
-import { taskList } from '../../../testUtils/mockApi/syncTaskList/data';
-
 import EmitterKey from '../../../data/EmitterKey';
 import EventEmitter from '../../../utils/EventEmitter';
-
 import UpdateSyncTask from '.';
 import { createSpySuccessResponse } from '@actiontech/shared/lib/testUtil/mockApi';
+import { DataSourceManagerSegmentedKey } from '../../DataSourceManagement/index.type';
+import { getBySelector } from '@actiontech/shared/lib/testUtil/customQuery';
+import { syncTaskDetailMockData } from '../../../testUtils/mockApi/syncTaskList/data';
 
 jest.mock('react-router-dom', () => {
   return {
@@ -23,7 +20,6 @@ jest.mock('react-router-dom', () => {
 
 describe('page/SyncDataSource/UpdateSyncTask', () => {
   const navigateSpy = jest.fn();
-  const projectID = mockProjectInfo.projectID;
   const taskId = '1739531854064652288';
 
   const customRender = () => {
@@ -38,7 +34,6 @@ describe('page/SyncDataSource/UpdateSyncTask', () => {
     });
     syncTaskList.mockAllApi();
     ruleTemplate.mockAllApi();
-    mockUseCurrentProject();
   });
 
   afterEach(() => {
@@ -49,13 +44,7 @@ describe('page/SyncDataSource/UpdateSyncTask', () => {
 
   it('render edit database snap', async () => {
     const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(1000));
-
-    expect(screen.getByText('返回同步任务列表')).toBeInTheDocument();
-    expect(screen.getByText('编辑同步任务')).toBeInTheDocument();
-    expect(screen.getByText('提 交')).toBeInTheDocument();
-    expect(screen.getByText('基础配置')).toBeInTheDocument();
-    expect(screen.getByText('SQL审核配置')).toBeInTheDocument();
+    await act(async () => jest.advanceTimersByTime(3000));
 
     expect(baseElement).toMatchSnapshot();
   });
@@ -77,18 +66,17 @@ describe('page/SyncDataSource/UpdateSyncTask', () => {
     const eventEmitSpy = jest.spyOn(EventEmitter, 'emit');
     const requestDetail = syncTaskList.getTaskSource();
     const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(300));
     expect(baseElement).toMatchSnapshot();
     await act(async () => jest.advanceTimersByTime(3000));
     expect(requestDetail).toHaveBeenCalled();
     expect(requestDetail).toHaveBeenCalledWith({
-      database_source_service_uid: taskId,
-      project_uid: projectID
+      db_service_sync_task_uid: taskId
     });
     expect(baseElement).toMatchSnapshot();
 
     fireEvent.click(screen.getByText('重 置'));
-    await act(async () => jest.advanceTimersByTime(300));
+    await act(async () => jest.advanceTimersByTime(0));
+
     expect(eventEmitSpy).toHaveBeenCalledWith(
       EmitterKey.DMS_SYNC_TASK_RESET_FORM
     );
@@ -97,31 +85,57 @@ describe('page/SyncDataSource/UpdateSyncTask', () => {
   it('render update task submit', async () => {
     const requestDetail = syncTaskList.getTaskSource();
     const requestUpdate = syncTaskList.updateTaskSource();
-    const detailInfo = { ...taskList[0] };
-    delete detailInfo.uid;
-    delete detailInfo.last_sync_success_time;
-    delete detailInfo.project_uid;
-    customRender();
-    await act(async () => jest.advanceTimersByTime(3300));
+    const { container } = customRender();
+
     expect(requestDetail).toHaveBeenCalled();
+    await act(async () => jest.advanceTimersByTime(3300));
+
+    expect(container).toMatchSnapshot();
+
+    fireEvent.change(getBySelector('#url', container), {
+      target: {
+        value: 'http://192.168.1.3:3333'
+      }
+    });
+    fireEvent.change(getBySelector('#params_key1', container), {
+      target: { value: 'params__1' }
+    });
+    fireEvent.change(getBySelector('#params_key2', container), {
+      target: { value: 432 }
+    });
+    fireEvent.click(getBySelector('#needAuditForSqlQuery', container));
+
     fireEvent.click(screen.getByText('提 交'));
-    await act(async () => jest.advanceTimersByTime(300));
+    await act(async () => jest.advanceTimersByTime(0));
     expect(screen.getByText('提 交').parentNode).toHaveClass('ant-btn-loading');
     expect(screen.getByText('重 置').parentNode).toHaveAttribute('disabled');
     await act(async () => jest.advanceTimersByTime(3000));
     expect(requestUpdate).toHaveBeenCalled();
     expect(requestUpdate).toHaveBeenCalledWith({
-      database_source_service: {
-        ...detailInfo
-      },
-      database_source_service_uid: taskId,
-      project_uid: projectID
+      db_service_sync_task_uid: taskId,
+      db_service_sync_task: {
+        name: syncTaskDetailMockData.name,
+        source: syncTaskDetailMockData.source,
+        db_type: syncTaskDetailMockData.db_type,
+        cron_express: syncTaskDetailMockData.cron_express,
+        url: 'http://192.168.1.3:3333',
+        sqle_config: {
+          ...syncTaskDetailMockData.sqle_config,
+          sql_query_config: {
+            audit_enabled: false
+          }
+        },
+        additional_params: [
+          { key: 'key1', value: 'params__1' },
+          { key: 'key2', value: '432' }
+        ]
+      }
     });
     expect(screen.getByText('同步任务编辑成功')).toBeInTheDocument();
     await act(async () => jest.advanceTimersByTime(3000));
     expect(navigateSpy).toHaveBeenCalled();
     expect(navigateSpy).toHaveBeenCalledWith(
-      `/project/${projectID}/sync-data-source`,
+      `/data-source-management?active=${DataSourceManagerSegmentedKey.SyncDataSource}`,
       {
         replace: true
       }
@@ -137,13 +151,11 @@ describe('page/SyncDataSource/UpdateSyncTask', () => {
       })
     );
     const { baseElement } = customRender();
-    await act(async () => jest.advanceTimersByTime(300));
     expect(baseElement).toMatchSnapshot();
     await act(async () => jest.advanceTimersByTime(3000));
     expect(requestDetail).toHaveBeenCalled();
     expect(requestDetail).toHaveBeenCalledWith({
-      database_source_service_uid: taskId,
-      project_uid: projectID
+      db_service_sync_task_uid: taskId
     });
     expect(baseElement).toMatchSnapshot();
   });
