+++ packages/base/src/page/SyncDataSource/UpdatePage/index.tsx
@@ -11,7 +11,7 @@ import {
 } from '@actiontech/shared';
 import { PageLayoutHasFixedHeaderStyleWrapper } from '@actiontech/shared/lib/styleWrapper/element';
 import { useForm } from 'antd/es/form/Form';
-import SyncTaskForm, { SyncTaskFormFields } from '../Form';
+import SyncTaskForm from '../Form';
 import { ResponseCode } from '@actiontech/shared/lib/enum';
 import { IGetDBServiceSyncTask } from '@actiontech/shared/lib/api/base/service/common';
 import EmitterKey from '../../../data/EmitterKey';
@@ -23,6 +23,7 @@ import DBServiceSyncTaskService from '@actiontech/shared/lib/api/base/service/DB
 import { IUpdateDBServiceSyncTaskParams } from '@actiontech/shared/lib/api/base/service/DBServiceSyncTask/index.d';
 import useTaskSource from '../../../hooks/useTaskSource';
 import useAsyncParams from 'sqle/src/components/BackendForm/useAsyncParams';
+import { SyncTaskFormFields } from '../Form/index.type';
 
 const UpdateSyncTask: React.FC = () => {
   const { t } = useTranslation();
@@ -45,32 +46,34 @@ const UpdateSyncTask: React.FC = () => {
     const values: SyncTaskFormFields = await form.validateFields();
     startSubmit();
 
-    const formParams = taskSourceTips.generateTaskSourceAdditionalParams(
+    const additionalParams = taskSourceTips.generateTaskSourceAdditionalParams(
       values.source
     );
 
     const params: IUpdateDBServiceSyncTaskParams = {
       db_service_sync_task_uid: taskId ?? '',
-      name: values.name,
-      source: values.source,
-      db_type: values.instanceType,
-      // #if [sqle]
-      sqle_config: {
-        rule_template_id: values.ruleTemplateId,
-        rule_template_name: values.ruleTemplateName,
-        sql_query_config: {
-          audit_enabled: values.needAuditForSqlQuery,
-          allow_query_when_less_than_audit_level:
-            values.allowQueryWhenLessThanAuditLevel
-        }
-      },
-      // #endif
-      cron_express: values.syncInterval,
-      url: values.url,
-      additional_params: mergeFromValueIntoParams(
-        values.params,
-        formParams ?? []
-      )
+      db_service_sync_task: {
+        name: values.name,
+        source: values.source,
+        db_type: values.instanceType,
+        // #if [sqle]
+        sqle_config: {
+          rule_template_id: values.ruleTemplateId,
+          rule_template_name: values.ruleTemplateName,
+          sql_query_config: {
+            audit_enabled: values.needAuditForSqlQuery,
+            allow_query_when_less_than_audit_level:
+              values.allowQueryWhenLessThanAuditLevel
+          }
+        },
+        // #endif
+        cron_express: values.syncInterval,
+        url: values.url,
+        additional_params: mergeFromValueIntoParams(
+          values.params,
+          additionalParams ?? []
+        )
+      }
     };
     DBServiceSyncTaskService.UpdateDBServiceSyncTask(params)
       .then((res) => {
